refactor(commands): extract remote-control commands into remote.rs
